Amazon-AWS-Modify-Instance-Attribute2 - reimplement UpdateInstanceType op into ModifyInstanceAttribute op using QueryAPI pattern - moved up one level QueryApiExecutor 2. Signed off by Tusa Mihai: tusa.mihai@gmail.com
