+++ cs-amazon/src/main/java/io/cloudslang/content/amazon/entities/aws/ImageState.java
@@ -12,7 +12,7 @@ public enum ImageState {
     PENDING,
     FAILED;
 
-    public static String getValue(String input) throws Exception {
+    public static String getValue(String input) {
         if (isBlank(input)) {
             return NOT_RELEVANT;
         }
