+++ cs-amazon/src/main/java/io/cloudslang/content/amazon/entities/aws/ImageType.java
@@ -12,7 +12,7 @@ public enum ImageType {
     KERNEL,
     RAMDISK;
 
-    public static String getValue(String input) throws Exception {
+    public static String getValue(String input) {
         if (isBlank(input)) {
             return NOT_RELEVANT;
         }
