+++ cs-amazon/src/main/java/io/cloudslang/content/amazon/entities/aws/NetworkInterfaceAttachmentStatus.java
@@ -13,7 +13,7 @@ public enum NetworkInterfaceAttachmentStatus {
     DETACHING,
     DETACHED;
 
-    public static String getValue(String input) throws Exception {
+    public static String getValue(String input) {
         if (isBlank(input)) {
             return NOT_RELEVANT;
         }
