+++ cs-amazon/src/main/java/io/cloudslang/content/amazon/entities/aws/VirtualizationType.java
@@ -11,7 +11,7 @@ public enum VirtualizationType {
     PARAVIRTUAL,
     HVM;
 
-    public static String getValue(String input) throws Exception {
+    public static String getValue(String input) {
         if (isBlank(input)) {
             return NOT_RELEVANT;
         }
