+++ cs-amazon/src/main/java/io/cloudslang/content/amazon/entities/aws/VolumeType.java
@@ -14,7 +14,7 @@ public enum VolumeType {
     SC1,
     ST1;
 
-    public static String getValue(String input) throws Exception {
+    public static String getValue(String input) {
         if (isBlank(input)) {
             return NOT_RELEVANT;
         }
