+++ cs-amazon/src/main/java/io/cloudslang/content/amazon/factory/ComputeFactory.java
@@ -13,7 +13,7 @@ import io.cloudslang.content.amazon.services.impl.OpenstackComputeServiceImpl;
 public class ComputeFactory {
     private static final String AMAZON = "amazon";
 
-    public static ComputeService getComputeService(CommonInputs commonInputs) throws Exception {
+    public static ComputeService getComputeService(CommonInputs commonInputs) {
         ComputeService computeService;
         switch (commonInputs.getProvider().toLowerCase()) {
             case AMAZON:
