+++ cs-amazon/src/main/java/io/cloudslang/content/amazon/utils/InputsUtil.java
@@ -288,7 +288,7 @@ public final class InputsUtil {
         return input;
     }
 
-    public static String getValidEbsSize(String input, String ebsType) throws Exception {
+    public static String getValidEbsSize(String input, String ebsType) {
         if (NOT_RELEVANT.equalsIgnoreCase(input)) {
             return NOT_RELEVANT;
         }
